# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MVP multi-tenant event platform. Auth owns user identity; org membership and permissions live in organization/iam.",
        "title": "Rallya API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/orgs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Orgs the caller belongs to, with roles. Used for org switching and GET /auth/me bootstrap.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orgs"
                ],
                "summary": "List my organizations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/orgdto.Org"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a tenant org; the caller becomes OWNER.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orgs"
                ],
                "summary": "Create organization",
                "parameters": [
                    {
                        "description": "Org payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/orgdto.CreateOrg"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/orgdto.Org"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "409": {
                        "description": "Slug taken",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/orgs/invites/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Caller account email must match the invite. Creates the membership; idempotent for members.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orgs"
                ],
                "summary": "Accept invite",
                "parameters": [
                    {
                        "description": "Invite token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/orgdto.AcceptInvite"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orgdto.Org"
                        }
                    },
                    "400": {
                        "description": "Invalid/expired or email mismatch",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/orgs/invites/decline": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Rejects a pending invite addressed to the caller. Idempotent; accepted invites cannot be declined.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orgs"
                ],
                "summary": "Decline invite",
                "parameters": [
                    {
                        "description": "Invite token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/orgdto.DeclineInvite"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invite declined",
                        "schema": {
                            "$ref": "#/definitions/orgdto.MessageAlias"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "409": {
                        "description": "Already accepted",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/orgs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Org details; :id accepts UUID or slug. Non-members get 404.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orgs"
                ],
                "summary": "Get organization",
                "parameters": [
                    {
                        "type": "string",
                        "example": "acme",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orgdto.Org"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Hard delete; memberships and invites cascade.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orgs"
                ],
                "summary": "Delete organization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orgdto.MessageAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Rename / re-slug / logo. Slug conflicts return 409.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orgs"
                ],
                "summary": "Update organization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/orgdto.UpdateOrg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orgdto.Org"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    }
                }
            }
        },
...
        "/orgs/{id}/invites/{inviteId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orgs"
                ],
                "summary": "Revoke invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Invite UUID",
                        "name": "inviteId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orgdto.MessageAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/orgdto.ErrorAlias"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "orgdto.AcceptInvite": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "minLength": 1,
                    "example": "d9e4f3a2b1c04852e5f6a708192b3c4d5e6f8091a2b3c4d5e6f708192b3"
                }
            }
        },
...
        "orgdto.CreateOrg": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "logo": {
                    "type": "string",
                    "maxLength": 2048,
                    "example": "https://example.com/logo.png"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Acme Inc"
                },
                "slug": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "acme"
                }
            }
        },
...
        "orgdto.ErrorAlias": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "insufficient role"
                }
            }
        },
...
        "orgdto.Org": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-09-19T12:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "2b3c4d5e-6f70-8a9b-0c1d-2e3f4a5b6c7d"
                },
                "logoUrl": {
                    "type": "string",
                    "example": "https://example.com/logo.png"
                },
                "name": {
                    "type": "string",
                    "example": "Acme Inc"
                },
                "role": {
                    "type": "string",
                    "example": "OWNER"
                },
                "slug": {
                    "type": "string",
                    "example": "acme"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003caccessToken\u003e from POST /auth/login (or /auth/refresh).\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}